Further attempts to scale graphics

+++ block_game.py
@@ -14,7 +14,7 @@ pygame.font.init()
 ########################################################################################################################
 # Display setup
 
-SIZE = pygame.display.list_modes()[0][1] / 2
+SIZE = int(pygame.display.list_modes()[0][1] / 2)
 screen = pygame.display.set_mode((SIZE, SIZE))
 pygame.display.set_caption('''Block Game''')
 pygame.display.set_icon(pygame.image.load('icon.png'))
@@ -22,7 +22,7 @@ pygame.mouse.set_visible(True)
 
 ########################################################################################################################
 # Global Variables
-BLOCK_SIZE = SIZE * 0.0375
+BLOCK_SIZE = int(SIZE * 0.0375)
 PIECE_NAMES = pieces.PIECE_NAMES
 
 ########################################################################################################################
